+++ ChangeLog.md
@@ -5,7 +5,11 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 # TO-DO List
 
 - Add command `/log copy <player>` or `/logCopy <player>`, to copy logger configurations from other players.
-- Modify command `/tick warp` with new options: `interrupt`, `status`, etc.
-- Add HUD logger `tickWarp` with options: `bar`, `value`.
+- Add field `NAME` for all the logHelpers.
+- Display the actual option when setting logger with default options.
+    - `Subscribed to <logger> \(<option>\)`
 
-# Features
+# Feature Modifications
+
+- Add HUD logger `tickWarp` with options: `bar`, `value`.
+- Add options `interrupt` and `status` for command `/tick warp`.